+++ packages/substrate-bindings/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- `ethAccount` correctly encodes valid addresses.
+
 ### Added
 
 - Export `Keccak256`